remove after_import_batch_validation from model/program and into a validator class, make duplicate_facility_description_results more Rails-y
